style: Req/Res Dto 스타일 통일

+++ server/server/src/main/java/com/mjubus/server/exception/TaxiParty/TaxiPartyNotFoundException.java
@@ -1,6 +1,5 @@
 package com.mjubus.server.exception.TaxiParty;
 
-import com.mjubus.server.domain.TaxiParty;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
@@ -8,7 +7,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 public class TaxiPartyNotFoundException extends RuntimeException{
     private String taxipartyId;
 
-    public TaxiPartyNotFoundException(Integer id) {
+    public TaxiPartyNotFoundException(Long id) {
         super(id.toString());
 
         taxipartyId = id.toString();
